Add US Daily Data Page
This page is based heavily on the states data page but·
surfaces the US Daily numbers.

I tried to keep basically the exact same styling as the states·
data page. I think it looks a little less good in this context, but·
I thought consistency was better than trying to tweak it for this.

This page can be navigated to by following a link at the top of the
data page.

I don't think this is ideal but seemed less bad than other options.
A footer link seemed wrong since we have a huge "TAKE ME TO THE DATA" button
that seems like the canonical way to get to the data, and there is already
a link in the footer that is called "The Data". It felt weird and wrong
to me to have another link to a different kind of data in the same set
of links.
I thought about putting the data on the same page as the main data page
but it felt more organized and right to have it on its own separate page.

I think it would be reasonable to put the US Daily Data above the state
level data on the data page, but that change might be surprising or
confusing to existing users so I decided that this solution was the
best tradeoff.

For future work, I think a graph showing change over time could·
be compelling but there is no effort at that as part of this change.

I tested this manually on my local machine.