Differentiate Last updated from Last refreshed
We don't actually "update" any data when we refresh/pull from the API. Which means
the timestamp listed falsly suggest we update the data show on the States and
US Daily pages when `sheets.updated` changes.

I think "refresh" avoids that potentially confusing term but it does leave
some ambiguity which I think could either be solved with copy on the page or
using an info tooltip for people who care what "Last refreshed" means exactly.