Use relative links when appropriate
There's no need for us to hardcode the domain for any path that falls under
covidtracker.com inside of our Markdown. It adds visual clutter for no good
reason.

The only slight advantage on the API page is that the API links work on
localhost development if they're pointing to covidtracker.com and they won't
work with this change since you might not be running the API and routing it
appropriately.

Still, I don't think maybe people hit the API endpoints from their local
machine.