fix(visualization): Changed tooltip style for map
Note that the positioning of the tooltip was also changed in this commit. Previously, the tooltip was `position: fixed`, which caused the tooltip to not scroll with the page. Switched it to absolute + added a new relative div to position it consistently across the choropleth and bubble maps.

fixes issue #353