fix(dashboard): normalize CDC and Covid Tracking data for comparison chartsi
Fixes COVID19Tracking/website#530

The issue here was the number of items in the two data arrays was entirely
mismatched. And the scaleBand x-axis in BarChart merely dumps a bar for each
item in the array.

Switching to scaleTime would not be appropriate for maintaining the bar chart
look and feel (it is better for plotting line graphs).

Approach here is to add any missing date items (with null values) to each of
the arrays. After normalization, they will have equal number of items, which
in turn makes the x-ticks and scale comparable between the two charts.