feat(search): extract Search logic to context module
The search feature was started as a PoC using State hooks,
that was extracted to src/context/search-context,
using Context hooks for easier reuse (on /search but also in the header).

Added withSearch HOC, which wraps withLocation, to re-use
searchContext easily accross the site while retaining access to URL query.

All logic specific to search, but not to search page, moved to
the hook file.

Pages that match have their (full, for now) body rendered with
some highlighting (deserves some more emphasis to make it more visible).
If the body has no hightlighting we return its truncated value.

Finally - more like a side effect, but a nice one - the search is
triggered whenever the input value is updated. We need to check:
1. Is this going to consume our API quota too fast, and
2. Will this be, OK as the header will be an autocomplete,
and this input refreshes the result page instantly? The UX feels right
but maybe I'm just happy that it works.