Update download.js
To me "Download all available data for a single state" button indicates that button itself is a download link, not an accordion button.  The skinny down arrow looks to much like a line or a letter to be much help (atleast in Win10).

Hopefully "Open listing of individual states' data to download" is more clear to most people.

Thanks for all your hard work!